Add algorithm for solving reef coral task
Module is divided into preprocessing part and the one with the main logic and functionality. determine_coral_reef_changes is used to combine all logic and preprocessing parts. The key to solve it was to get white and pink parts of the coral reef and then work on them (using bitwise operations). We needed to align two pictures to have similar perspective (align_photos function).